+++ apps/core-demo/src/app/app.module.ts
@@ -1,15 +1,36 @@
 import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterModule } from '@angular/router';
+import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { AppComponent } from './app.component';
+import { TargetBlankDirective } from './directives/target-blank.directive';
+import { MaterialModule } from './modules/material.module';
+import { JwtDemoPageComponent } from './pages/jwt-demo-page.component';
+import { TypedocPageComponent } from './pages/typedoc-page.component';
 
 @NgModule({
-	declarations: [AppComponent],
+	declarations: [AppComponent, TypedocPageComponent, TargetBlankDirective],
 	imports: [
 		BrowserModule,
+		BrowserAnimationsModule,
 		HttpClientModule,
-		RouterModule.forRoot([], { initialNavigation: 'enabled' }),
+		MaterialModule,
+		FontAwesomeModule,
+		RouterModule.forRoot(
+			[
+				{
+					path: 'typedoc',
+					component: TypedocPageComponent,
+				},
+				{
+					path: 'jwt',
+					component: JwtDemoPageComponent,
+				},
+			],
+			{ initialNavigation: 'enabled', relativeLinkResolution: 'legacy' }
+		),
 	],
 	providers: [],
 	bootstrap: [AppComponent],