+++ libs/jwt/src/index.ts
@@ -5,4 +5,6 @@ export * from './lib/model/auth-core-configuration.interface';
 export * from './lib/model/header-configuration.interface';
 export { HttpMethod } from './lib/model/http-method.enum';
 export { JwtToken, JwtTokenPair, JwtTokenPayload } from './lib/model/jwt-token.class';
+export * from './lib/providers/create-jwt-configuration-provider.function';
+export * from './lib/providers/create-jwt-refresh-configuration-provider.function';
 export { JwtTokenService } from './lib/service/jwt-token.service';