+++ libs/jwt/src/lib/function/http-response.predicate.ts
@@ -1,5 +1,4 @@
 import { HttpEvent, HttpEventType, HttpResponse } from '@angular/common/http';
 
-export function isHttpResponse<T>(httpEvent: HttpEvent<T>): httpEvent is HttpResponse<T> {
-	return httpEvent.type === HttpEventType.Response;
-}
+export const isHttpResponse = <T>(httpEvent: HttpEvent<T>): httpEvent is HttpResponse<T> =>
+	httpEvent.type === HttpEventType.Response;