+++ libs/jwt/src/lib/function/is-not-nullish.predicate.ts
@@ -1,3 +1,1 @@
-export function isNotNullish<T>(t: T | undefined | null): t is T {
-	return t !== undefined && t !== null;
-}
+export const isNotNullish = <T>(t: T | undefined | null): t is T => t !== undefined && t !== null;