+++ libs/jwt/src/lib/function/match-against.function.ts
@@ -1,9 +1,13 @@
 import { isString } from './string.predicate';
 
 export const matchRule = (rule: string | RegExp, against?: string | null): boolean => {
-	if (isString(rule)) return rule === against;
-	else if (against) return rule.test(against);
-	else return false;
+	if (isString(rule)) {
+		return rule === against;
+	} else if (against) {
+		return rule.test(against);
+	} else {
+		return false;
+	}
 };
 
 /**