+++ libs/jwt/src/lib/providers/create-jwt-configuration-provider.function.ts
@@ -1,0 +1,19 @@
+import {
+	JwtConfigurationProvider,
+	JwtModuleConfigurationProvider,
+	JWT_CONFIGURATION_TOKEN,
+} from '../token/jwt-configuration.token';
+
+/**
+ * Helps you define a JwtConfigurationProvider
+ *
+ * @internal
+ */
+export const createJwtConfigurationProvider = (
+	tokenConfigurationProvider: JwtModuleConfigurationProvider
+): JwtConfigurationProvider =>
+	({
+		provide: JWT_CONFIGURATION_TOKEN,
+		multi: false,
+		...tokenConfigurationProvider,
+	} as JwtConfigurationProvider);