+++ libs/ngrx/jest.config.js
@@ -1,16 +1,16 @@
 module.exports = {
 	name: 'ngrx',
-	preset: '../../jest.config.js',
+	preset: '../../jest.preset.js',
 	coverageDirectory: '../../coverage/libs/ngrx',
 	snapshotSerializers: [
-		'jest-preset-angular/build/AngularNoNgAttributesSnapshotSerializer.js',
-		'jest-preset-angular/build/AngularSnapshotSerializer.js',
-		'jest-preset-angular/build/HTMLCommentSerializer.js',
+		'jest-preset-angular/build/serializers/no-ng-attributes',
+		'jest-preset-angular/build/serializers/ng-snapshot',
+		'jest-preset-angular/build/serializers/html-comment',
 	],
 	setupFilesAfterEnv: ['<rootDir>/src/test-setup.ts'],
 	globals: {
 		'ts-jest': {
-			tsConfig: '<rootDir>/tsconfig.spec.json',
+			tsconfig: '<rootDir>/tsconfig.spec.json',
 			stringifyContentPathRegex: '\\.(html|svg)$',
 			astTransformers: {
 				before: [